Game starts on click
