+++ index.html
@@ -6,7 +6,10 @@
   <title>Helicopter Time</title>
 </head>
 <body>
-  <canvas id="game" class="game-canvas" width="800px" height="600px"></canvas>
+  <div class="copter-game">
+    <h2 id="start-game">Click to begin</h2>
+    <canvas id="game" class="game-canvas" width="800px" height="600px"></canvas>
+  </div>
   <script src="https://code.jquery.com/jquery-2.2.3.min.js"   integrity="sha256-a23g1Nt4dtEYOj7bR+vTu7+T8VP13humZFBJNIYoEJo="   crossorigin="anonymous"></script>
   <script src="main.bundle.js"></script>
 </body>
